Prevention of FilesActivity being killed by reading whether this is a configuration change.
Downside:
May not get the correct value and the expected behaviour may not exist in pre API-11 devices.

But those market shares...
